Rename *Reader files and fix Data*TaskTest
