+++ tests/src/Pipeline/Task/DataUpsertTaskTest.php
@@ -12,7 +12,7 @@ use Liquetsoft\Fias\Component\Pipeline\Task\Task;
 use Liquetsoft\Fias\Component\Serializer\FiasSerializer;
 use Liquetsoft\Fias\Component\Storage\Storage;
 use Liquetsoft\Fias\Component\Tests\BaseCase;
-use Liquetsoft\Fias\Component\XmlReader\BaseXmlReader;
+use Liquetsoft\Fias\Component\Reader\BaseReader;
 
 /**
  * Тест для задачи, которая обновляет данные данные из файла в БД.
@@ -49,7 +49,7 @@ class DataUpsertTaskTest extends BaseCase
         $state = new ArrayState;
         $state->setParameter(Task::FILES_TO_INSERT_PARAM, [__DIR__ . '/_fixtures/data.xml']);
 
-        $task = new DataUpsertTask($entityManager, new BaseXmlReader, $storage, new FiasSerializer);
+        $task = new DataUpsertTask($entityManager, new BaseReader, $storage, new FiasSerializer);
         $task->run($state);
 
         $this->assertSame([321], $insertedData);
